FEAT-#139; REFACTOR-#100: Implement the spawning of MPI processes dynamically
