+++ unidist/config/__init__.py
@@ -13,6 +13,7 @@ from .backends.ray import (
     RayObjectStoreMemory,
 )
 from .backends.dask import DaskMemoryLimit, IsDaskCluster, DaskSchedulerAddress
+from .backends.mpi import IsMpiSpawnWorkers, MpiHosts
 from .parameter import ValueSource
 
 __all__ = [
@@ -26,5 +27,7 @@ __all__ = [
     "DaskMemoryLimit",
     "IsDaskCluster",
     "DaskSchedulerAddress",
+    "IsMpiSpawnWorkers",
+    "MpiHosts",
     "ValueSource",
 ]
