+++ unidist/core/backends/mpi/core/worker/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2021-2022 Modin authors
+#
+# SPDX-License-Identifier: Apache-2.0
